+++ CHANGES.md
@@ -1,4 +1,5 @@
 [unreleased]
+* add `afragen/add-plugin-dependency-api` as composer requirement
 * update test plugins removing `hello-dolly` and adding `git-updater` as non dot org example
 
 #### 1.4.1 / 2022-08-18